+++ futurehome/src/ha/publish_device.ts
@@ -54,6 +54,7 @@ import { sensor_watpressure__components } from '../services/sensor_watpressure';
 import { sensor_wattemp__components } from '../services/sensor_wattemp';
 import { sensor_weight__components } from '../services/sensor_weight';
 import { thermostat__components } from '../services/thermostat';
+import { water_heater__components } from '../services/water_heater';
 import { abbreviateHaMqttKeys } from './abbreviate_ha_mqtt_keys';
 import { ha } from './globals';
 import { HaMqttComponent } from './mqtt_components/_component';
@@ -209,6 +210,7 @@ const serviceHandlers: {
   sensor_wattemp: sensor_wattemp__components,
   sensor_weight: sensor_weight__components,
   thermostat: thermostat__components,
+  water_heater: water_heater__components,
 };
 
 export function haPublishDevice(parameters: {